Refactor for new landing page
The new subscriptions app is a simple landing page. As such, this
commit removes the old templates and greatly simplifies the app.

A new landing page template has been created but the designs are not
final so the cut-up hasn't started beyond broadly adding the main
sections so there's a real page there to check.